Add a basic implementation using AiiDA
Specifically `aiida-shell`. Note that this implementation intentionally
provides the most simple approach. Not all steps are captured in
provenance and the entire procedure is not wrapped into a single
workflow. Also the execution of the QE calculations are not parallelized
but are performed sequential. All of these steps could easily be
improved, but it is not clear that these are required.